requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pydantic>=2.0.0
PyYAML>=6.0
gspread>=5.0.0
google-auth>=2.0.0
schedule>=1.2.0
python-dotenv>=1.0.0 